fixed some new errors regarding delete posts and removed unused code from files

+++ src/main/java/nl/garagemeijer/salesapi/services/BusinessVehicleService.java
@@ -48,6 +48,9 @@ public class BusinessVehicleService {
     }
 
     public void deleteBusinessVehicle(Long id) {
+        if (businessVehicleRepository.findById(id).isEmpty()) {
+            throw new RecordNotFoundException("Business vehicle with id: " + id + " not found");
+        }
         businessVehicleRepository.deleteById(id);
     }
 }
